feat: update brewfile
python is not updated as it is managed using mise.
vim is not updated as I don't use it and it carries a few dependencies for nothing.
commitizen carries too many dependencies and it's a python package, I prefer native packages.
added parallel and noti as I now use them frequently in CLI.